#60 Remove redundancy in fm_url's
Main approach is to move many of the url parts to consants in constants.py.

Prepare for / extend on new sensor_id's for new energy providers.

Util_functions.py renamed to v2g_globals.py because it also contains initialisation code for CONSTANTS. These init-constants ultimately should be moved to this module but i'v not been able to so they remain in te constants.py.